Revise implementation
